Accepted :)))))

+++ src/WeightedSuccessor.java
@@ -2,9 +2,9 @@ public class WeightedSuccessor {
 
     final byte i;
     final byte j;
-    final byte weight;
+    final int weight;
 
-    public WeightedSuccessor(byte i, byte j, byte weight) {
+    public WeightedSuccessor(byte i, byte j, int weight) {
         this.i = i;
         this.j = j;
         this.weight = weight;
